Added some new classes for 3D and organized some folders
