+++ SDL_OpenGL_Platformer.vcxproj.filters
@@ -336,10 +336,6 @@
     <ClInclude Include="src\components\Components.h">
       <Filter>src\ECS\Components</Filter>
     </ClInclude>
-    <ClInclude Include="src\components\RigidBodyComponent.h" />
-    <ClInclude Include="src\components\SpriteComponent.h">
-      <Filter>src\ECS\Components</Filter>
-    </ClInclude>
     <ClInclude Include="src\graphics\TextureManager.h">
       <Filter>src\ECS</Filter>
     </ClInclude>
@@ -355,7 +351,6 @@
     <ClInclude Include="src\components\Transform.h">
       <Filter>src\ECS\Components</Filter>
     </ClInclude>
-    <ClInclude Include="src\components\Collider.h" />
     <ClInclude Include="src\graphics\Shader.h" />
     <ClInclude Include="src\core\GameManager.h">
       <Filter>src</Filter>
@@ -364,11 +359,17 @@
       <Filter>src</Filter>
     </ClInclude>
     <ClInclude Include="src\custom\Player.h" />
-    <ClInclude Include="src\core\Physics.h" />
     <ClInclude Include="src\tiles\Tilemap.h" />
     <ClInclude Include="src\tiles\Tile.h" />
     <ClInclude Include="src\custom\Camera.h" />
-    <ClInclude Include="src\components\AnimComponent.h" />
+    <ClInclude Include="src\components\3D\RigidBody3D.h" />
+    <ClInclude Include="src\components\2D\RigidBodyComponent.h" />
+    <ClInclude Include="src\components\2D\SpriteComponent.h">
+      <Filter>src\ECS\Components</Filter>
+    </ClInclude>
+    <ClInclude Include="src\components\Collider.h" />
+    <ClInclude Include="src\core\2D\Physics2D.h" />
+    <ClInclude Include="src\core\3D\Physics3D.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\math\VMath.cpp">
@@ -381,7 +382,6 @@
       <Filter>src\math</Filter>
     </ClCompile>
     <ClCompile Include="src\components\ECS.cpp" />
-    <ClCompile Include="src\components\RigidBodyComponent.cpp" />
     <ClCompile Include="src\graphics\TextureManager.cpp">
       <Filter>src\ECS</Filter>
     </ClCompile>
@@ -393,7 +393,6 @@
     </ClCompile>
     <ClCompile Include="src\components\Transform.cpp" />
     <ClCompile Include="src\graphics\Shader.cpp" />
-    <ClCompile Include="src\components\SpriteComponent.cpp" />
     <ClCompile Include="src\core\GameManager.cpp">
       <Filter>src</Filter>
     </ClCompile>
@@ -404,11 +403,14 @@
       <Filter>src</Filter>
     </ClCompile>
     <ClCompile Include="src\custom\Player.cpp" />
-    <ClCompile Include="src\core\Physics.cpp" />
     <ClCompile Include="src\tiles\Tilemap.cpp" />
     <ClCompile Include="src\tiles\Tile.cpp" />
     <ClCompile Include="src\custom\Camera.cpp" />
-    <ClCompile Include="src\components\AnimComponent.cpp" />
+    <ClCompile Include="src\components\3D\RigidBody3D.cpp" />
+    <ClCompile Include="src\components\2D\RigidBodyComponent.cpp" />
+    <ClCompile Include="src\components\2D\SpriteComponent.cpp" />
+    <ClCompile Include="src\core\2D\Physics2D.cpp" />
+    <ClCompile Include="src\core\3D\Physics3D.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="external\dll\glew32.dll">
