--- src/components/AnimComponent.cpp
@@ -1,1 +1,0 @@
-#include "AnimComponent.h"
