--- src/components/AnimComponent.h
@@ -1,11 +1,0 @@
-#ifndef ANIMCOMPONENT_H
-#define ANIMCOMPONENT_H
-
-
-class AnimComponent
-{
-
-private:
-public:
-};
-#endif
