+++ src/components/Components.h
@@ -1,5 +1,5 @@
 //This file should have includes to separate components
 //If a component includes this file we'll run into circular dependency problems
-#include "SpriteComponent.h"
-#include "RigidBodyComponent.h"
+#include "2D/SpriteComponent.h"
+#include "2D/RigidBodyComponent.h"
 #include "Transform.h"
