+++ src/core/3D/Physics3D.cpp
@@ -1,0 +1,7 @@
+#include "Physics3D.h"
+#include "tiles/Tile.h"
+#include "math/Vector.h"
+#include "math/VMath.h"
+
+using namespace MATH;
+
