+++ src/core/3D/Physics3D.h
@@ -1,0 +1,20 @@
+#ifndef PHYSICS3D_H
+#define PHYSICS3D_H
+
+#include "math/VMath.h"
+#include "MMath.h"
+#include "components/3D/RigidBody3D.h"
+
+class Physics2D
+{
+private:
+
+	
+
+	
+public:
+	
+
+};
+
+#endif
